src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Joaquin Huaman Pino<jhuamanp22@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.15.2"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
model-bakery = "^1.20.3"
drf-yasg = "^1.21.8"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.4"
pylint-django = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
